Feature: stat info & optimize

+++ src/AwakenServer.Application/StatInfo/StatInfoOptions.cs
@@ -5,7 +5,7 @@ namespace AwakenServer.StatInfo;
 public class StatInfoOptions
 {
     public List<int> Periods { get; set; }
-    public Dictionary<string, long> TypePeriodMapping { get; set; }
+    public Dictionary<string, int> TypePeriodMapping { get; set; }
     public List<string> StableCoinPriority { set; get; } = new();
     public string DataVersion { get; set; }
 }
